refactor(domain): 로또들의 결과를 맵으로 반환하는 메서드를 다시 추가
Map 타입을 반환해야 하는 것이 못마땅하여 삭제했던 메서드이지만
당첨 등수와 대응하는 당첨 개수의 관리는 Map으로 관리하는 것이 적절한데
해당 Map을 만들어줄 객체는 Lottos밖에 없다고 판단하여 삭제했던 메서드를
다시 부활시킴
