+++ src/main/java/lotto/utils/PrintingMessages.java
@@ -1,0 +1,2 @@
+package lotto.utils;public class PrintingMessages {
+}
